tests/wpt/tests/css/css-device-adapt/viewport-user-scalable-no-clamp-to-max.tentative.html

<!doctype html>
<meta charset=utf-8>
<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=device-width, maximum-scale=1.0">
<link rel="help" href="https://drafts.csswg.org/css-device-adapt/">
<link rel="help" href="https://webcompat.com/issues/52856">
<style>
body {
  margin: 0;
}
#content {
  width: 100px;
  height: 100px;
  background: green;
}
</style>

<div id="content">Content</div>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
'use strict';
test(() => {
  assert_equals(window.visualViewport.scale, 1.0,
    'visual viewport scale should be 1.0');
}, 'Page with meta viewport "width=device-width, user-scalable=no, ' +
   'initial-scale=device-width, maximum-scale=1.0" ' +
   'should scale to 1.0.');
</script>
